personal/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.0
Faker==21.0.0
